+++ src/properties/logger.ts
@@ -137,8 +137,18 @@ export default {
     `Found ${num} private dns zones`,
   // Public Ips
   foundPublicIps: (num: number): string => `Found ${num} public ips`,
+  // Recovery Vaults
+  foundRecoveryVaults: (num: number): string => `Found ${num} recovery vaults`,
   // RedisCache
   foundRedisCaches: (num: number): string => `Found ${num} Redis caches`,
+  // Replication Appliances
+  foundReplicationAppliances: (num: number): string => `Found ${num} replication appliances`,
+  // Replication Centers
+  foundReplicationCenters: (num: number): string => `Found ${num} replication centers`,
+  // Replication Networks
+  foundReplicationNetworks: (num: number): string => `Found ${num} replication networks`,
+  // Replication Policies
+  foundReplicationPolicies: (num: number): string => `Found ${num} replication policies`,
   // Resource Groups
   foundResourceGroups: (num: number): string => `Found ${num} resource groups`,
   // Security Assesments