+++ src/services/replicationAppliances/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazureReplicationApplianceInput!]!) {
+  addazureReplicationAppliance(input: $input, upsert: true) {
+    numUids
+  }
+}`;