+++ src/services/replicationPolicies/connections.ts
@@ -3,30 +3,25 @@ import { isEmpty } from 'lodash'
 
 import services from '../../enums/services'
 import { caseInsensitiveEqual } from '../../utils'
-import { RawAzureLogAnalyticsWorkspace } from '../logAnalyticsWorkspace/data'
 import { RawAzureResourceGroup } from '../resourceGroup/data'
-import { RawAzureLogAnalyticsSolution } from './data'
+import { RawAzureReplicationPolicy } from './data'
 
 export default ({
   service,
   data,
   region,
 }: {
-  service: RawAzureLogAnalyticsSolution
+  service: RawAzureReplicationPolicy
   data: Array<{ name: string; data: { [property: string]: any[] } }>
   region: string
 }): {
   [property: string]: ServiceConnection[]
 } => {
   const connections: ServiceConnection[] = []
-  const {
-    id,
-    resourceGroupId: rgName,
-    properties: { workspaceResourceId } = {},
-  } = service
+  const { id, resourceGroupId: rgName } = service
 
   /**
-   * Find resource group related to this log analytics solutions instance
+   * Find resource group related to this replication policy
    */
   const resourceGroups: {
     name: string
@@ -47,33 +42,6 @@ export default ({
           resourceType: services.resourceGroup,
           relation: 'child',
           field: 'resourceGroup',
-        })
-      }
-    }
-  }
-
-  /**
-   * Find log analytics workspace related to this log analytics solutions instance
-   */
-  const workspaces: {
-    name: string
-    data: { [property: string]: RawAzureLogAnalyticsWorkspace[] }
-  } = data.find(({ name }) => name === services.logAnalyticsWorkspace)
-
-  if (workspaces?.data?.[region]) {
-    const workspacesInRegion: RawAzureLogAnalyticsWorkspace[] = workspaces.data[
-      region
-    ].filter(({ id: logAnalyticsWorkspaceId }: RawAzureLogAnalyticsWorkspace) =>
-      caseInsensitiveEqual(logAnalyticsWorkspaceId, workspaceResourceId)
-    )
-
-    if (!isEmpty(workspacesInRegion)) {
-      for (const rg of workspacesInRegion) {
-        connections.push({
-          id: rg.id,
-          resourceType: services.logAnalyticsWorkspace,
-          relation: 'child',
-          field: 'logAnalyticsWorkspace',
         })
       }
     }