+++ src/services/replicationPolicies/mutation.ts
@@ -1,0 +1,5 @@
+export default `mutation($input: [AddazureReplicationPolicyInput!]!) {
+  addazureReplicationPolicy(input: $input, upsert: true) {
+    numUids
+  }
+}`;